Move GSD to hoomd.write
