+++ hoomd/__init__.py
@@ -19,16 +19,17 @@ import sys
 import os
 
 from hoomd import version
+from hoomd import trigger
+from hoomd import variant
+from hoomd import filter
+from hoomd import device
 from hoomd import analyze
+from hoomd import update
+from hoomd import integrate
 from hoomd import benchmark
 from hoomd import comm
-from hoomd import dump
-from hoomd import integrate
-from hoomd import update
 from hoomd import util
-from hoomd import variant
-from hoomd import device
-from hoomd import trigger
+from hoomd import write
 try:
     from hoomd import md
 except ImportError:
@@ -53,7 +54,6 @@ from hoomd.state import State
 from hoomd.operations import Operations
 from hoomd.snapshot import Snapshot
 from hoomd import tune
-from hoomd import output
 from hoomd import logging
 from hoomd import custom
