+++ hoomd/write/__init__.py
@@ -1,1 +1,2 @@
+from hoomd.write.gsd import GSD
 from hoomd.write.csv_backend import CSV
